internationalization_from_adminpanel/lang_app/requirements.txt
Django==1.11.15 
django-filter==2.0.0 
django-modeltranslation==0.12.2 
djangorestframework==3.8.2 
Markdown==2.6.11 
pytz==2018.5